split up big analysis file
